Feature widget_download is now its own tab
